AVSfldIO.jl Documentation

Contents

AVSfldIO.jl Documentation

Contents

Overview

Index

Functions

Overview

File I/O routines for AVS .fld format data files; see (AVSfldIO)

Index

AVSfldIO.AVSfldIO

AVSfldIO.arg_get

AVSfldIO.datatype_fld_to_mat

AVSfldIO.fld_header

AVSfldIO.fld_open

AVSfldIO.fld_read

AVSfldIO.fld_write

AVSfldIO.fld_write_data_fix

AVSfldIO.string_to_array

Functions

AVSfldIO.AVSfldIO — Module

AVSfldIO

module for AVS .fld file IO

AVSfldIO.arg_get — Function

arg_get(head, name, toint)

Parse an argument from header, of the name=value form

AVSfldIO.datatype_fld_to_mat — Method

format, endian, bytes = datatype_fld_to_mat(datatype)

Determine data format from .fld header datatype.

AVSfldIO.fld_header — Method

head = fld_header(file::String ; dir::String="", chat=false)

Read header data from AVS format .fld file, then close file.

in

file::String file name, usually ending in .fld

option

dir::String prepend file name with this directory; default ""

chat::Bool verbose? default: false

out

head::String array of header information

AVSfldIO.fld_open — Method

head, is_external_file, fid = fld_open(file ; dir::String="", chat=false)

Read header data from AVS format .fld file. Leaves file open for more reading.

in

file::String file name, usually ending in .fld

option

dir::String prepend file name with this directory; default ""

chat::Bool verbose? default: false

out

head::String array of header information

is_external_file::Bool true if AVS external format

fid::IOstream

AVSfldIO.fld_read — Method

data = fld_read(file::String ; dir::String="", chat=false)

Read data from AVS format .fld file

in

file file name, usually ending in .fld

option

dir::String prepend file name with this directory; default ""

chat::Bool verbose?

out

data Array (1D - 5D) in the data type of the file itself

AVSfldIO.fld_write — Method

fld_write(file, data ; kwargs...)

Write data into AVS format .fld file. See README for file format.

in

file name of file typically ending in .fld

data real data array

option

check::Bool report error if file exists; default true

dir::String directory name to prepend file name; default ""

endian::Symbol:lelittle endian (default),:be` big endian

head::Array{String} comment information for file header

raw::Bool put raw data in name.raw, header in name.fld where file = name.fld; default false

AVSfldIO.fld_write_data_fix — Method

data = fld_write_data_fix(data)

Convert data to format suitable for writing to .fld file.

AVSfldIO.string_to_array — Method

header = string_to_array(header_lines)

Convert long string with embedded newlines into string array.